Implement camera modes with secondary axis
LOCKED remains locked to the vessel, but FREE and CHASE use the
in-vector as primary reference with velocity as secondary, and direction
to target as secondary for chase. Orbital uses velocity as primary and
in as secondary (might change this, uncertain, as it makes EVA kerbals a
little odd, but then again, TINU).