Implement numpad view switching
Pressing numpad 7, 1, 3 or 9 gives orthogonal control camera switching.
 - 7: dorsal view with fore pointing camera-up
 - 1: aft view with dorsal pointing camera-up
 - 3: starboard view with dorsal pointing camera-up
 - 9: rotate 180 degrees around camera-horizontal axis
Holding the right control key modifies the above:
 - rctrl-7: ventral view with fore pointing camera-down
 - rctrl-1: fore view with ventral pointing camera-down
 - rctrl-3: port view with ventral pointing camera-down
 - rctrl-9: rotate 180 degrees around camera-vertical axis
This should make TINU much more usable when docking

The descriptions above are for Locked camera mode. Other modes currently
use the EVA frame of reference, but are subject to change in the future.