Add a key to toggle the TINU panel and TINU itself
The numpad-/ key will toggle the control panel giving access even when
the game is paused and the main UI is disabled (UI hiding is respected
when not paused). Right-control-/ toggles TINU itself for the current
camera mode (ensuring disable-all is false). ie, no matter how TINU is
disabled rctrl-/ will always enable TINU in the current mode, or disable
TINU for only the current mode.