Make target direction primary for orbital chase camera
This makes chase camera work fairly consistently regardless
of the relative positions of vessel and target, and the orbital velocity
vector (as it is now irrelevant).

The target direction is still secondary (with radial as primary) for
"free" (non-orbital) case camera, keeping atmospheric and ground
operations the same as always.

In addition, the reference vectors are always updated when the mode is
changed, making for less weird camera angle snaps.